Fixed a sneaky bug resulting in crashes loading presets.
Former-commit-id: 59ee161171f27c467c31eabd420147cf7dcb7fb1

+++ Source/MainEditor.h
@@ -128,7 +128,7 @@
 			// state variables.
 			int refreshRate;
 			int viewTopCoord;
-			std::size_t selTab;
+			std::int32_t selTab;
 			cpl::iCtrlPrec_t oldRefreshRate;
 			bool unFocused, idleInBack, isEditorVisible, firstKioskMode, hasAnyTabBeenSelected;
 			juce::Point<int> kioskCoords;
